+++ common-api/src/main/java/cn/featherfly/common/app/Platform.java
@@ -1,0 +1,14 @@
+
+package cn.featherfly.common.app;
+
+/**
+ * Platform.
+ *
+ * @author zhongj
+ */
+public interface Platform {
+
+    int id();
+
+    String name();
+}